backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
PyJWT==2.8.0
argon2-cffi==23.1.0
pytest==8.2.0
requests==2.31.0
